refactor: atom placement; attributes form; remove cache
